requirements.txt
Cython
librosa
matplotlib
numpy
phonemizer
scipy
tensorboard
onnx
onnxruntime
.DS_Store
epub2txt
mlx
datasets
tqdm
requests
soundfile
opencv-python
uvicorn
torch
safetensors
tokenizers
pillow
typer
rich
lightning